Create enumerate
This is in C++ - so i'll just put it here as one idea for how the algorithm might work. I don't remember exactly why I had to create the two separate functions, but I think that javascript's looser variable types will make only one necessary.

I gave this up when I started to realize that I would need to identify 1)unpunctuated numbers 2)numbers separated by commas (e.g.1,000,000) 3) decimals (either 1,000.14 or .14 or 0.14) - then I realized that it would have to be the opposite for European numbers, and gave up.

But creating a settings page is the perfect solution! Users can just choose whether they should convert numbers as American or European.

I didn't create the converter function for decimals yet - but the simplest thing it seems to me would be to search for the relevant punctuation mark (period or comma) and then just transliterate as "point number number number number" or "comma number number number" - but if you'd like to get fancier (fourteen hundredths, etc.) go for it!